safari-600.7-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.4_Comments/S7.4_A2_T2-expected.txt

CONSOLE MESSAGE: line 83: SyntaxError: Multiline comment was not closed properly
S7.4_A2_T2

PASS Expected parsing failure

TEST COMPLETE
